nuri : 검색조건 추가

+++ src/main/java/com/example/demo/portfolio/vo/PortfolioVO.java
@@ -6,7 +6,7 @@ import java.util.Date;
 import java.util.List;
 
 public class PortfolioVO {
-    private int portNo;
+    private Long portNo;
     private String userId;
     private int projectNo;
     @DateTimeFormat(pattern = "yyyy-mm-dd")
@@ -25,11 +25,11 @@ public class PortfolioVO {
     public PortfolioVO() {
     }
 
-    public int getPortNo() {
+    public Long getPortNo() {
         return portNo;
     }
 
-    public void setPortNo(int portNo) {
+    public void setPortNo(Long portNo) {
         this.portNo = portNo;
     }
